backend/requirements.txt
# Web framework
flask==3.0.3
flask-cors==4.0.1

# AWS SDK
boto3==1.34.144
botocore==1.34.144

# Database
PyMySQL==1.1.1
SQLAlchemy==2.0.31

# Auth & Security
PyJWT==2.8.0
cryptography==42.0.8

# Validation
marshmallow==3.21.3
email-validator==2.2.0

# Environment
python-dotenv==1.0.1

# WSGI server (production)
gunicorn==22.0.0

# Utilities
shortuuid==1.0.13
python-dateutil==2.9.0